MOVEONLY: Move address related functions from rpcwallet to addresses.cpp
